+++ assignment-001/src/helperfile.c
@@ -1,21 +1,21 @@
 #include"../include/headerfile.h"
 
-int AddNums(int iNum1, int iNum2)
+float AddNums(float iNum1, float iNum2)
 {
   return (iNum1 + iNum2);
 }
 
-int SubNums(int iNum1, int iNum2)
+float SubNums(float iNum1, float iNum2)
 {
   return (iNum1 - iNum2);
 }
 
-int MultNums(int iNum1, int iNum2)
+float MultNums(float iNum1, float iNum2)
 {
   return (iNum1 * iNum2);
 }
 
-float DivNums(int iNum1, int iNum2)
+float DivNums(float iNum1, float iNum2)
 {
   return (iNum1 / (float)iNum2);
 }
